Drop tables if process.env.DROP_TABLES=true
This will prevent the accidental dropping of the users,projects tables

+++ migrations/20131122023554-make-user-table.js
@@ -16,5 +16,9 @@ exports.up = function(db, callback) {
 };
 
 exports.down = function(db, callback) {
-  db.dropTable('users', callback);
+  if (process.env.DROP_TABLES == "true") {
+    db.dropTable('users', callback);
+  } else {
+    callback();
+  }
 };
