+++ migrations/20131123161823-add-project-model.js
@@ -14,5 +14,9 @@ exports.up = function(db, callback) {
 };
 
 exports.down = function(db, callback) {
-  db.dropTable('projects', callback);
+  if (process.env.DROP_TABLES == "true") {
+    db.dropTable('projects', callback);
+  } else {
+    callback();
+  }
 };
